short => byte, SLIM-534

+++ osgp-dto/src/main/java/com/alliander/osgp/dto/valueobjects/smartmetering/ObisCodeValuesDto.java
@@ -13,14 +13,14 @@ public class ObisCodeValuesDto implements Serializable {
 
     private static final long serialVersionUID = 6725616781596815362L;
 
-    private short a;
-    private short b;
-    private short c;
-    private short d;
-    private short e;
-    private short f;
+    private byte a;
+    private byte b;
+    private byte c;
+    private byte d;
+    private byte e;
+    private byte f;
 
-    public ObisCodeValuesDto(final short a, final short b, final short c, final short d, final short e, final short f) {
+    public ObisCodeValuesDto(final byte a, final byte b, final byte c, final byte d, final byte e, final byte f) {
         super();
         this.a = a;
         this.b = b;
@@ -30,27 +30,27 @@ public class ObisCodeValuesDto implements Serializable {
         this.f = f;
     }
 
-    public short getA() {
+    public byte getA() {
         return this.a;
     }
 
-    public short getB() {
+    public byte getB() {
         return this.b;
     }
 
-    public short getC() {
+    public byte getC() {
         return this.c;
     }
 
-    public short getD() {
+    public byte getD() {
         return this.d;
     }
 
-    public short getE() {
+    public byte getE() {
         return this.e;
     }
 
-    public short getF() {
+    public byte getF() {
         return this.f;
     }
 }
